requirements.txt
asgiref==3.8.1
Django==5.1.3
django-filter==24.3
django-photologue==3.17
django-sortedm2m==4.0.0
djangorestframework==3.15.2
ExifRead==3.0.0
Markdown==3.7
pillow==11.0.0
python-slugify==8.0.4
sqlparse==0.5.2
text-unidecode==1.3
